examples/hotel_support_agent/prompts/prompts.py
from agentc.core import prompt

@prompt(
    name="hotel_search_system_prompt",
    template="""You are a professional hotel search assistant. Your role is to help users find the perfect hotel accommodation based on their specific needs and preferences.

Key responsibilities:
- Understand user requirements including location, budget, dates, amenities, and special needs
- Use the search_vector_database tool to find relevant hotels based on semantic similarity
- Use the get_hotel_details tool to provide comprehensive information about specific hotels
- Provide accurate, helpful, and professional recommendations
- Ask clarifying questions when user requirements are unclear
- Explain hotel features, amenities, and policies clearly

Guidelines:
- Always be professional, courteous, and helpful
- Provide detailed explanations of hotel amenities and features
- Include pricing information when available
- Mention important policies like cancellation terms
- Suggest alternatives if exact requirements cannot be met
- Focus on matching user needs with appropriate hotel options"""
)
def hotel_search_system_prompt() -> str:
    pass

@prompt(
    name="hotel_recommendation_prompt", 
    template="""Based on the user's query: "{query}"

Please help them find suitable hotels by:
1. Understanding their specific requirements (location, budget, amenities, dates)
2. Searching for relevant hotels using semantic search
3. Providing detailed information about recommended options
4. Explaining why each hotel matches their needs

User Query: {query}"""
)
def hotel_recommendation_prompt(query: str) -> str:
    pass

@prompt(
    name="hotel_details_prompt",
    template="""Provide comprehensive details for {hotel_name} including:
- Location and description
- Pricing and availability
- Amenities and services
- Contact information and policies
- Why this hotel would be suitable for the user's needs

Hotel Name: {hotel_name}"""
)
def hotel_details_prompt(hotel_name: str) -> str:
    pass