examples/hotel_support_agent/pyproject.toml
[tool.poetry]
name = "hotel-support-agent"
version = "0.1.0"
description = "A hotel support agent using Couchbase."
authors = ["Your Name <you@example.com>"]

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.0"
langgraph = "^0.1.0"
agentc-core = {path = "../../agent-catalog/libs/agentc_core", develop = true}
agentc-langgraph = {path = "../../agent-catalog/libs/agentc_integrations/langgraph", develop = true}
couchbase = "^4.0.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
langchain-couchbase = "^0.3.0"
langchain-openai = "^0.3.13"
datasets = "^3.5.0"
getpass4 = "^0.0.14"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"